requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.6.15
channels==4.2.2
charset-normalizer==3.4.2
django==4.2.21
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
joblib==1.4.2
numpy==1.24.4
pandas==2.0.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
